pyproject.toml
[project]
name = "projectfromnull"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.17.1",
    "dotenv>=0.9.9",
    "mypy>=1.20.2",
    "ruff>=0.15.12",
]
